+++ attributes/default.rb
@@ -28,6 +28,7 @@ default['kubernetes']['kubeconfig']                         = '/etc/kubernetes/k
 default['kubernetes']['tls_cert_file']                      = '/etc/kubernetes/ssl/apiserver.pem'
 default['kubernetes']['tls_private_key_file']               = '/etc/kubernetes/ssl/apiserver-key.pem'
 default['kubernetes']['client_ca_file']                     = '/etc/kubernetes/ssl/ca.pem'
+default['kubernetes']['requestheader_client_ca_file']       = '/etc/kubernetes/ssl/ca.pem'
 default['kubernetes']['service_account_key_file']           = '/etc/kubernetes/ssl/ca-key.pem'
 default['kubernetes']['cluster_signing_cert_file']          = '/etc/kubernetes/ssl/ca.pem'
 default['kubernetes']['cluster_signing_key_file']           = '/etc/kubernetes/ssl/ca-key.pem'