+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# UNRELEASED
+- ApiServer: added `--requestheader-client-ca-file` cmdline flag
+
 # 1.11.6 (26.07.2018)
 - Etcd: 3.3.9
 - Weave: 2.4.0