+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [Kubernetes: 1.11.2](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.11.md#changelog-since-v1111)
 - [cri-o: 1.11.2](https://github.com/kubernetes-incubator/cri-o/releases/tag/v1.11.2)
 - [buildah: 1.3](https://github.com/projectatomic/buildah/releases/tag/v1.3)
+- [podman: 0.8.3](https://github.com/containers/libpod/releases/tag/v0.8.3)
 
 # 1.11.6 (26.07.2018)
 - Etcd: 3.3.9